regions: Separate utility functions.
